Notification systems working
Converted a lot of files to tab indenting.
Wrapped the notification messaging in a service.
Notifications are now logged in the database.
Created a security service to give us information about who is logged
in.
Created a security controller to render the information about who is
logged in.
Updated the toolbar to accept incoming notifications and update the UI.
Notifications are now queried out on page load (for persistence)
Cleaned out some commented code.